replace LifecycleLogger with CurrentScreenManager
